Merge upstream changes

+++ gst/nnstreamer/include/nnstreamer_plugin_api_trainer.h
@@ -33,11 +33,11 @@ typedef struct _GstTensorTrainerProperties
   GstTensorsInfo input_meta;    /**< configured input tensor info */
   const char *model_config;    /**< The configuration file path for creating model */
   const char *model_save_path;    /**< The file path to save the model */
-  int64_t num_inputs;    /**< The number of input lists, the input is where framework receive the features to train the model, num_inputs indicates how many inputs there are. */
-  int64_t num_labels;    /**< The number of label lists, the label is where framework receive the class to train the model, num_labels indicates how many labels there are. */
-  int64_t num_training_samples;    /**< The number of training sample used to train the model. */
-  int64_t num_validation_samples;    /**< The number of validation sample used to valid the model. */
-  int64_t num_epochs;    /**< The number of repetition of total training and validation sample. subplugin must receive total samples((num_training_samples + num_validation_samples) * num_epochs) */
+  unsigned int num_inputs;    /**< The number of input lists, the input is where framework receive the features to train the model, num_inputs indicates how many inputs there are. */
+  unsigned int num_labels;    /**< The number of label lists, the label is where framework receive the class to train the model, num_labels indicates how many labels there are. */
+  unsigned int num_training_samples;    /**< The number of training sample used to train the model. */
+  unsigned int num_validation_samples;    /**< The number of validation sample used to valid the model. */
+  unsigned int num_epochs;    /**< The number of repetition of total training and validation sample. subplugin must receive total samples((num_training_samples + num_validation_samples) * num_epochs) */
 
   GCond *training_complete_cond;    /**< Tensor trainer wait when receive EOS before model training is complete, subplugin should send signal when model training is complete. */
 } GstTensorTrainerProperties;
@@ -50,8 +50,8 @@ typedef struct _GstTensorTrainerProperties
 typedef struct _GstTensorTrainerFrameworkInfo
 {
   const char *name;    /**< Name of the neural network framework, searchable by FRAMEWORK property. */
-  int is_training_complete;  /**< Check if training is complete, Use int instead of gboolean because this is refered by custom plugins. */
-  int64_t epoch_cnt;    /**< Number of currently completed epochs */
+  unsigned int is_training_complete;  /**< Check if training is complete, Use unsigned int instead of gboolean because this is refered by custom plugins. */
+  unsigned int epoch_cnt;    /**< Number of currently completed epochs */
 } GstTensorTrainerFrameworkInfo;
 
 typedef struct _GstTensorTrainerFramework GstTensorTrainerFramework;
